requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3.post1
pywatchman==1.4.1
sqlparse==0.4.4
